empire/x-analysis/requirements.txt
redis[hiredis]==5.2.1
anthropic==0.42.0
chromadb==0.5.23
httpx==0.28.1
python-dotenv==1.0.1
yt-dlp==2024.12.23
